watermark_photos/api/requirements.txt
flask==2.3.3
pillow==10.0.0
gunicorn==21.2.0
requests==2.31.0
python-telegram-bot==20.7
